# lib/api-zod/src/generated/types/activitiesPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rift Youth Leadership Network API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface ActivitiesPage {
  data: Activity[];
  total: number;
  page: number;
  limit: number;
}