# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rift Youth Leadership Network API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityInputStatus } from './activityInputStatus';

export interface ActivityInput {
  /** @minLength 1 */
  title: string;
  description: string;
  date: string;
  location: string;
  status: ActivityInputStatus;
  imageUrl?: string;
}