# lib/api-zod/src/generated/types/activityUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rift Youth Leadership Network API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityUpdateStatus } from './activityUpdateStatus';

export interface ActivityUpdate {
  title?: string;
  description?: string;
  date?: string;
  location?: string;
  status?: ActivityUpdateStatus;
  imageUrl?: string;
}