# lib/api-zod/src/generated/types/galleryImageUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rift Youth Leadership Network API
 * OpenAPI spec version: 0.1.0
 */
import type { GalleryImageUpdateCategory } from './galleryImageUpdateCategory';

export interface GalleryImageUpdate {
  title?: string;
  imageUrl?: string;
  category?: GalleryImageUpdateCategory;
}