# lib/api-zod/src/generated/types/leadershipInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rift Youth Leadership Network API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadershipInput {
  name: string;
  position: string;
  bio: string;
  photoUrl?: string;
  sortOrder?: number;
}